+++ photutils/aperture/mask.py
@@ -2,6 +2,7 @@
 """
 This module defines a class for aperture masks.
 """
+import warnings
 
 import astropy.units as u
 import numpy as np
@@ -38,7 +39,6 @@ class ApertureMask:
         """
         Array representation of the mask data array (e.g., for matplotlib).
         """
-
         return self.data
 
     @property
@@ -46,71 +46,34 @@ class ApertureMask:
         """
         The shape of the mask data array.
         """
-
         return self.data.shape
 
-    def _overlap_slices(self, shape):
+    def get_overlap_slices(self, shape):
         """
-        Calculate the slices for the overlapping part of the bounding
-        box and an array of the given shape.
+        Get slices for the overlapping part of the aperture mask and a
+        2D array.
 
         Parameters
         ----------
-        shape : tuple of int
-            The ``(ny, nx)`` shape of array where the slices are to be
-            applied.
+        shape : 2-tuple of int
+            The shape of the 2D array.
 
         Returns
         -------
-        slices_large : tuple of slices
+        slices_large : tuple of slices or `None`
             A tuple of slice objects for each axis of the large array,
             such that ``large_array[slices_large]`` extracts the region
             of the large array that overlaps with the small array.
-
-        slices_small : slice
-            A tuple of slice objects for each axis of the small array,
-            such that ``small_array[slices_small]`` extracts the region
-            of the small array that is inside the large array.
+            `None` is returned if there is no overlap of the bounding
+            box with the given image shape.
+        slices_small : tuple of slices or `None`
+            A tuple of slice objects for each axis of the aperture mask
+            array such that ``small_array[slices_small]`` extracts the
+            region that is inside the large array. `None` is returned if
+            there is no overlap of the bounding box with the given image
+            shape.
         """
-
-        if len(shape) != 2:
-            raise ValueError('input shape must have 2 elements.')
-
-        xmin = self.bbox.ixmin
-        xmax = self.bbox.ixmax
-        ymin = self.bbox.iymin
-        ymax = self.bbox.iymax
-
-        if xmin >= shape[1] or ymin >= shape[0] or xmax <= 0 or ymax <= 0:
-            # no overlap of the aperture with the data
-            return None, None
-
-        slices_large = (slice(max(ymin, 0), min(ymax, shape[0])),
-                        slice(max(xmin, 0), min(xmax, shape[1])))
-
-        slices_small = (slice(max(-ymin, 0),
-                              min(ymax - ymin, shape[0] - ymin)),
-                        slice(max(-xmin, 0),
-                              min(xmax - xmin, shape[1] - xmin)))
-
-        return slices_large, slices_small
-
-    def _to_image_partial_overlap(self, image):
-        """
-        Return an image of the mask in a 2D array, where the mask
-        is not fully within the image (i.e., partial or no overlap).
-        """
-
-        # find the overlap of the mask on the output image shape
-        slices_large, slices_small = self._overlap_slices(image.shape)
-
-        if slices_small is None:
-            return None  # no overlap
-
-        # insert the mask into the output image
-        image[slices_large] = self.data[slices_small]
-
-        return image
+        return self.bbox.get_overlap_slices(shape)
 
     def to_image(self, shape):
         """
@@ -127,20 +90,18 @@ class ApertureMask:
         result : `~numpy.ndarray`
             A 2D array of the mask.
         """
-
         if len(shape) != 2:
             raise ValueError('input shape must have 2 elements.')
 
+        # find the overlap of the mask on the output image shape
+        slices_large, slices_small = self.bbox.get_overlap_slices(shape)
+
+        if slices_small is None:
+            return None  # no overlap
+
+        # insert the mask into the output image
         image = np.zeros(shape)
-
-        if self.bbox.ixmin < 0 or self.bbox.iymin < 0:
-            return self._to_image_partial_overlap(image)
-
-        try:
-            image[self.bbox.slices] = self.data
-        except ValueError:  # partial or no overlap
-            image = self._to_image_partial_overlap(image)
-
+        image[slices_large] = self.data[slices_small]
         return image
 
     def cutout(self, data, fill_value=0., copy=False):
@@ -177,36 +138,36 @@ class ApertureMask:
             ``fill_value``.  `None` is returned if there is no overlap
             of the aperture with the input ``data``.
         """
-
         data = np.asanyarray(data)
         if data.ndim != 2:
             raise ValueError('data must be a 2D array.')
 
-        partial_overlap = False
-        if self.bbox.ixmin < 0 or self.bbox.iymin < 0:
-            partial_overlap = True
+        # find the overlap of the mask on the output image shape
+        slices_large, slices_small = self.bbox.get_overlap_slices(data.shape)
 
-        if not partial_overlap:
-            # try this for speed -- the result may still be a partial
-            # overlap, in which case the next block will be triggered
+        if slices_small is None:
+            return None  # no overlap
+
+        cutout_shape = (slices_small[0].stop - slices_small[0].start,
+                        slices_small[1].stop - slices_small[1].start)
+
+        if cutout_shape == self.shape:
+            cutout = data[slices_large]
             if copy:
-                cutout = data[self.bbox.slices].copy()  # preserves Quantity
-            else:
-                cutout = data[self.bbox.slices]
+                cutout = np.copy(cutout)
+            return cutout
 
-        if partial_overlap or (cutout.shape != self.shape):
-            slices_large, slices_small = self._overlap_slices(data.shape)
+        # cutout is always a copy for partial overlap
+        if ~np.isfinite(fill_value):
+            dtype = float
+        else:
+            dtype = data.dtype
+        cutout = np.zeros(self.shape, dtype=dtype)
+        cutout[:] = fill_value
+        cutout[slices_small] = data[slices_large]
 
-            if slices_small is None:
-                return None  # no overlap
-
-            # cutout is a copy
-            cutout = np.zeros(self.shape, dtype=data.dtype)
-            cutout[:] = fill_value
-            cutout[slices_small] = data[slices_large]
-
-            if isinstance(data, u.Quantity):
-                cutout = u.Quantity(cutout, unit=data.unit)
+        if isinstance(data, u.Quantity):
+            cutout <<= data.unit
 
         return cutout
 
@@ -236,15 +197,56 @@ class ApertureMask:
             is returned if there is no overlap of the aperture with the
             input ``data``.
         """
-
         cutout = self.cutout(data, fill_value=fill_value)
         if cutout is None:
             return None
         else:
             weighted_cutout = cutout * self.data
 
-            # needed to zero out non-finite data values outside of the
-            # mask but within the bounding box
-            weighted_cutout[self._mask] = 0.
+            # fill values outside of the mask but within the bounding box
+            weighted_cutout[self._mask] = fill_value
 
             return weighted_cutout
+
+    def get_values(self, data, mask=None):
+        """
+        Get the mask-weighted pixel values from the data as a 1D array.
+
+        If the ``ApertureMask`` was created with ``method='center'``,
+        (where the mask weights are only 1 or 0), then the returned
+        values will simply be pixel values extracted from the data.
+
+        Parameters
+        ----------
+        data : array_like or `~astropy.units.Quantity`
+            The 2D array from which to get mask-weighted values.
+
+        mask : array_like (bool), optional
+            A boolean mask with the same shape as ``data`` where a
+            `True` value indicates the corresponding element of ``data``
+            is not returned in the result.
+
+        Returns
+        -------
+        result : `~numpy.ndarray`
+            A 1D array of mask-weighted pixel values from the input
+            ``data``. If there is no overlap of the aperture with the
+            input ``data``, the result will be a 1-element array of
+            ``numpy.nan``.
+        """
+        slc_large, slc_small = self.bbox.get_overlap_slices(data.shape)
+        if slc_large is None:
+            return np.array([np.nan])
+        cutout = data[slc_large]
+        apermask = self.data[slc_small]
+        pixel_mask = (apermask > 0)  # good pixels
+
+        if mask is not None:
+            if mask.shape != data.shape:
+                raise ValueError('mask and data must have the same shape')
+            pixel_mask &= ~mask[slc_large]
+
+        # ignore multiplication with inf data values
+        with warnings.catch_warnings():
+            warnings.simplefilter('ignore', RuntimeWarning)
+            return (cutout * apermask)[pixel_mask]