+++ photutils/detection/__init__.py
@@ -4,5 +4,8 @@ This subpackage contains tools for detecting sources in an astronomical
 image.
 """
 
-from .core import *  # noqa
-from .findstars import *  # noqa
+from .base import *  # noqa
+from .daofinder import *  # noqa
+from .irafstarfinder import *  # noqa
+from .peakfinder import *  # noqa
+from .starfinder import *  # noqa