+++ photutils/morphology/core.py
@@ -35,13 +35,18 @@ def data_properties(data, mask=None, background=None):
 
     Returns
     -------
-    result : `~photutils.segmentation.SourceProperties` instance
-        A `~photutils.segmentation.SourceProperties` object.
+    result : `~photutils.segmentation.SourceCatalog` instance
+        A `~photutils.segmentation.SourceCatalog` object.
     """
+    # prevent circular imports
+    from ..segmentation import SegmentationImage, SourceCatalog
 
-    from ..segmentation import SourceProperties  # prevent circular imports
+    segment_image = SegmentationImage(np.ones(data.shape, dtype=int))
 
-    segment_image = np.ones(data.shape, dtype=int)
+    if background is not None:
+        background = np.atleast_1d(background)
+        if background.shape == (1,):
+            background = np.zeros(data.shape) + background
 
-    return SourceProperties(data, segment_image, label=1, mask=mask,
-                            background=background)
+    return SourceCatalog(data, segment_image, mask=mask,
+                         background=background)[0]